Application/requirements.txt
jsonschema
mysql-connector-python
tabulate
